fix betty 0, 1 and header

+++ 0-bubble_sort.c
@@ -9,7 +9,7 @@
 
 void bubble_sort(int *array, size_t size)
 {
-	long unsigned int i, j;
+	size_t i, j;
 	int tmp;
 
 	if (array == NULL || size == 0)
